pyproject.toml
[tool.poetry]
name = "ollamaexample"
version = "0.1.0"
description = ""
authors = ["fedenanni <nanni.federico@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
ollama = "^0.1.9"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
